\name{enaR-package}
\alias{enaR-package}
\alias{enaR}
\docType{package}
\title{
Tools ecological network analysis (ena) in R.
}
\description{
This package compiles functions for the analysis of ecological networks,
building on tools previously developed in the MatLab language (Fath and
Borrett 2006) with multiple additions of functionality. 
}
\details{
\tabular{ll}{
Package: \tab enaR \cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2012-10-10\cr
License: \tab GPL-2\cr
}
}
\author{
Authors: Stuart R. Borrett (borretts@uncw.edu) and Matthew K. Lau (mkl48@nau.edu)
Maintainer: Matthew K. Lau <mkl48@nau.edu>
}
\references{
Fath BD, Borrett SR. 2006. A Matlab Function for Network Environ
Analysis. Environmental Modelling and Software 21, 375-405.
}
\keyword{ package }
\seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% \code{\link[network:statnet-package]{statnet}}
%\code{\link[igraph:igraph-package]{igraph}}
\code{\link[network:network-package]{network}}
}

